requirements.txt
# Web Scraping & HTTP Requests
requests
beautifulsoup4
lxml

# Environment & Configuration
python-dotenv

# Logging
google-cloud-logging

# Web Framework & Server
Flask
gunicorn

# Email Notifications
sendgrid